image_processing_api/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
black==23.10.1
blinker==1.6.3
build==0.10.0
CacheControl==0.12.14
cachelib==0.9.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==2.0.1
click==8.1.7
contourpy==1.0.7
crashtest==0.4.1
cycler==0.11.0
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.7
Django==4.2.6
dnspython==2.4.2
dulwich==0.21.5
email-validator==2.1.0.post1
exceptiongroup==1.1.2
filelock==3.12.2
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
Flask-Login==0.6.2
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.39.3
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
installer==0.7.0
instaloader==4.9.5
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keyring==23.13.1
kiwisolver==1.4.4
lazy_loader==0.3
librosa==0.10.0.post2
limits==3.6.0
llvmlite==0.40.1
lockfile==0.12.2
loguru==0.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
more-itertools==10.1.0
msgpack==1.0.5
mypy-extensions==1.0.0
numba==0.57.1
numpy==1.24.2
opencv-python==4.8.1.78
ordered-set==4.1.0
packaging==23.0
panda==0.3.1
pandas==2.0.0
pathspec==0.11.2
pexpect==4.8.0
Pillow==9.5.0
pkginfo==1.9.6
platformdirs==3.10.0
pluggy==1.3.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
pooch==1.6.0
protobuf==4.21.4
ptyprocess==0.7.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
Pygments==2.16.1
PyMySQL==1.1.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
PyQt3D==5.15.5
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
rapidfuzz==2.15.1
redis==5.0.1
referencing==0.30.2
requests==2.28.1
requests-toolbelt==1.0.0
rich==13.6.0
rpds-py==0.9.2
scikit-learn==1.3.0
scipy==1.11.1
scrapfly-sdk==0.8.6
shellingham==1.5.3
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
SQLAlchemy==2.0.18
sqlparse==0.4.4
TBB==0.2
threadpoolctl==3.2.0
tomli==2.0.1
tomlkit==0.12.1
trove-classifiers==2023.8.7
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.13
virtualenv==20.24.3
webencodings==0.5.1
Werkzeug==2.3.7
wrapt==1.15.0
WTForms==3.1.0
xattr==0.10.1
zipp==3.16.2
